Fix 32bit builds (missing size_t string convert)
../libs/pbd/pbd/string_convert.h: In function 'std::string PBD::to_string(T) [with T = long unsigned int]':
../gtk2_ardour/foldback_strip.cc:581:   instantiated from here
../libs/pbd/pbd/string_convert.h:238: error: 'long unsigned int' is not a class, struc

+++ gtk2_ardour/foldback_strip.cc
@@ -565,7 +565,7 @@ FoldbackStrip::set_route (boost::shared_ptr<Route> rt)
 
 	_output_button.set_route (_route, this);
 
-	size_t number = 0;
+	int number = 0;
 	{
 		RouteList rl (_session->get_routelist (true, PresentationInfo::FoldbackBus));
 		RouteList::iterator i = find (rl.begin (), rl.end (), _route);
